pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "meemoo-sip-validator"
version = "0.1.0"
description = "A library that validates if a package is a valid meemoo SIP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-commons-ip>=0.0.1,<1",
    "xmlschema==4.1.0",
    "edtf-validate==2.0.0",
    "sipin-eark-models",
]
classifiers = [
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "build",
    "twine",
]

[tool.setuptools.package-data]
"meemoo_sip_validator" = ["assets/*.xml"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vv"